+++ svd-parser/src/readaction.rs
@@ -1,0 +1,14 @@
+use super::*;
+
+use crate::svd::ReadAction;
+impl Parse for ReadAction {
+    type Object = Self;
+    type Error = SVDErrorAt;
+    type Config = Config;
+
+    fn parse(tree: &Node, _config: &Self::Config) -> Result<Self, Self::Error> {
+        let text = tree.get_text()?;
+
+        Self::parse_str(text).ok_or_else(|| SVDError::InvalidReadAction(text.into()).at(tree.id()))
+    }
+}