+++ svd-rs/src/usage.rs
@@ -22,7 +22,7 @@ impl Default for Usage {
 }
 
 impl Usage {
-    /// Parse a string into an [Usage] value, returning [`Option::None`] if the string is not valid.
+    /// Parse a string into an [`Usage`] value, returning [`Option::None`] if the string is not valid.
     pub fn parse_str(s: &str) -> Option<Self> {
         match s {
             "read" => Some(Self::Read),